custom viewport and camera supported
fixed globalTimeglobalTime

+++ NPLRuntime/ParaEngine/3dengine/ParaXAnimInstance.cpp
@@ -24,7 +24,7 @@
 namespace ParaEngine
 {
 	// global time for global sequences
-	extern int globalTime;
+	extern int64_t globalTime;
 }
 
 using namespace ParaEngine;
